+++ opendut-lea/src/viper_tests/configurator/tabs/cluster.rs
@@ -1,6 +1,6 @@
 use leptos::prelude::*;
-use crate::tests::configurator::components::ClusterSelector;
-use crate::tests::configurator::types::UserTestConfiguration;
+use crate::viper_tests::configurator::components::ClusterSelector;
+use crate::viper_tests::configurator::types::UserTestConfiguration;
 
 #[component]
 pub fn ClusterTab(test_configuration: RwSignal<UserTestConfiguration>) -> impl IntoView {
