+++ opendut-lea/src/viper_tests/configurator/tabs/suite.rs
@@ -1,6 +1,6 @@
 use leptos::prelude::*;
-use crate::tests::configurator::components::TestSuiteSelector;
-use crate::tests::configurator::types::UserTestConfiguration;
+use crate::viper_tests::configurator::components::TestSuiteSelector;
+use crate::viper_tests::configurator::types::UserTestConfiguration;
 
 #[component]
 pub fn SuiteTab(test_configuration: RwSignal<UserTestConfiguration>) -> impl IntoView {
